+++ relay-server/src/utils/actix.rs
@@ -2,6 +2,9 @@ use ::actix::dev::{MessageResponse, ResponseChannel};
 use ::actix::fut::IntoActorFuture;
 use ::actix::prelude::*;
 use futures01::prelude::*;
+use tokio::runtime::Runtime;
+
+use relay_common::clone;
 
 pub enum Response<T, E> {
     Reply(Result<T, E>),
@@ -136,4 +139,23 @@ impl<T: 'static, E: 'static> Response<T, E> {
             Response::Future(future) => ActorResponse::Future(Box::new(fut::wrap_future(future))),
         }
     }
+}
+
+/// Constructs a single threaded tokio [`Runtime`] containing a clone of the actix [`System`].
+///
+/// This is required if you need to send messages from the tokio runtime to actix
+/// actors.
+///
+/// # Panics
+///
+/// The calling thread must have the actix system enabled, panics if this is invoked
+/// in a thread where actix is not enabled.
+pub fn tokio_runtime_with_actix() -> Runtime {
+    let system = System::current();
+    tokio::runtime::Builder::new_multi_thread()
+        .worker_threads(1)
+        .enable_all()
+        .on_thread_start(clone!(system, || System::set_current(system.clone())))
+        .build()
+        .unwrap()
 }